Move functions to correct file
